Added files via upload
CellNode class objects will function as vertices for the directed graph. Every time a Cell object is modified, it's corresponding CellNode should be updated and the in degrees as well. I was going to just use Cell as the vertex, but it would have been too much code in one class. Spreadsheet just has one method and the private 2-d array and graph. I'm trying to go by his assignment description.